umk: Fixed to compile
git-svn-id: svn://ultimatepp.org/upp/trunk@12626 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Draw/DrawTextUtil.cpp
@@ -102,7 +102,7 @@ void DrawTLText(Draw& draw, int x, int y, int cx, const wchar *text,
 				bool ak = false;
 				start = q;
 				while(q < s && *q >= ' ') {
-					if(akey && ToUpper(ToAscii(*q)) == akey && (apos == 0 || q - start + 1 == apos)) {
+					if(akey && (int)ToUpper(ToAscii(*q)) == akey && (apos == 0 || q - start + 1 == apos)) {
 						ak = true;
 						akey = 0;
 						break;
